+++ Neocom/Neocom/NCFittingShip.mm
@@ -245,7 +245,7 @@
 - (NSInteger) raceID {
 	NCVerifyFittingContext(self.engine);
 	auto ship = std::dynamic_pointer_cast<dgmpp::Ship>(self.item);
-	return ship ? ship->getRaceID() : 0;
+	return ship ? static_cast<NSInteger>(ship->getRaceID()) : 0;
 }
 
 - (double) capacity {
